backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
pydantic==2.9.2
psycopg2-binary==2.9.9
